Se finalizan los métodos para service completados

+++ core/Services/Services/ServicesService.cs
@@ -3,6 +3,7 @@ using core.Interfaces.Repositories.Services;
 using core.Interfaces.Services.IServicesService;
 using DTOs.Result;
 using DTOs.Service;
+using System.Reflection.PortableExecutable;
 using System.Security.Claims;
 
 namespace core.Services.Services
@@ -32,14 +33,63 @@ namespace core.Services.Services
                 }
                 filters.IdEnterprice = int.Parse(idEnterpriceClaim);
             }
+            else
+            {
+                if(idEnterprice == null)
+                {
+                    return ResultDto.FailResult("Al ser admin debes realizar la petición enviando el id de la empresa que deseas consultar", 401);
+                }
+            }
 
             var servicios = await _servicesRepository.GetServices(filters);
-            int statusCode = servicios.Any() ? 201 : 204;
+            if (servicios == null)
+            {
+                return ResultDto.FailResult("Ha sucedido un error al consultar", 401);
+            }
+            int statusCode = servicios.Any() ? 200 : 204;
             return ResultDto.SuccessResult(servicios, "Registros consultados exitosamente.", statusCode);
         }
         public async Task<ResultDto> UpdateServices(ServiceUpdateDto serviceUpdate, IEnumerable<Claim> claims)
         {
-            throw new NotImplementedException();
+            var roleClaim = claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+            var idEnterpriceClaim = claims.FirstOrDefault(c => c.Type == "idEnterprice")?.Value;
+            
+            if (string.IsNullOrEmpty(roleClaim))
+            {
+                return ResultDto.FailResult("Rol no encontrado en el token", 401);
+            }
+
+            ServiceFiltersDto filters = new ServiceFiltersDto { Id = serviceUpdate.Id };
+            if (roleClaim != "1")
+            {
+                if (string.IsNullOrEmpty(idEnterpriceClaim))
+                {
+                    return ResultDto.FailResult("El usuario no tiene una empresa asignada", 401);
+                }
+                filters.IdEnterprice = int.Parse(idEnterpriceClaim);
+            }
+            else
+            {
+                if (serviceUpdate.idEnterprice == 0)
+                {
+                    return ResultDto.FailResult("Debes enviar un idEnterprice como parametro", 400);
+                }
+                filters.IdEnterprice = serviceUpdate.idEnterprice;
+            }
+
+            var serviceFinded = await _servicesRepository.GetServices(filters);
+            if (serviceFinded == null || !serviceFinded.Any())
+            {
+                return ResultDto.FailResult("No se encontraron registros con ese id y tu id de empresa para actualizar", 400);
+            }
+
+            bool actualizado = await _servicesRepository.UpdateService(serviceUpdate);
+            if (!actualizado)
+            {
+                return ResultDto.FailResult("No se pudo actualizar el registro", 400);
+            }
+            return ResultDto.SuccessResult(message: "Registro actualizado correctamente", statusCode: 200);
+
         }
         public async Task<ResultDto> DeleteService(int id, IEnumerable<Claim> claims)
         {
@@ -62,16 +112,16 @@ namespace core.Services.Services
             }
 
             var serviceFinded = await _servicesRepository.GetServices(filters);
-            if (serviceFinded == null)
+            if (serviceFinded == null || !serviceFinded.Any())
             {
-                return ResultDto.FailResult("No se encontraron resultados para eliminar", 401);
+                return ResultDto.FailResult("No se encontraron registros con ese id para eliminar", 401);
             }
             bool eliminado = await _servicesRepository.DeleteService(id);
-            if (eliminado)
+            if (!eliminado)
             {
                 return ResultDto.FailResult("No se pudo eliminar el registro", 401);
             }
-                return ResultDto.SuccessResult(message: "Registro eliminado correctamente", statusCode: 200);
+            return ResultDto.SuccessResult(message: "Registro eliminado correctamente", statusCode: 200);
         }
         public async Task<ResultDto> InsertService(ServiceDto service, IEnumerable<Claim> claims)
         {
